// backup/stic_frontend/src/app/types/Materials.ts
export type Material = {
  id: number;
  name: string;
  density: number;
  pricePerKg: number;
  pricePerHourMachine: number;
  pricePerHourOperator: number;
};

export type MaterialPUT = {  
  name: string;
  density: number;
  pricePerKg: number;
  pricePerHourMachine: number;
  pricePerHourOperator: number;
};